Spatial DB
Implemented a public announcement System that would query a spatial
database. It kept track of all the areas the announcement system would
cover such as the buildings and the student location. The project
involved querying the database for retrieving the nearest neighbor,
finding intersection of regions, finding the surrounding regions and
plotting the values onto the location Map using JDBC.